massive cleanup and code revision
