+++ main.h
@@ -1,0 +1,10 @@
+// header guards
+#ifndef MAIN_H_
+#define MAIN_H_
+
+#include <stdio.h>
+#include <stdbool.h>
+#include <stdlib.h>
+#include "overflow.h"
+
+#endif
